Update Card component
- 크기가 작은 요소는 Card index.jsx 에 통합하여 관리하도록 함
- 재사용성을 고려하였을 때 재사용 가능성이 없는 컴포넌트들은 하나로 통합
