--- client/src/components/Card/PriceContainer/index.jsx
@@ -1,25 +1,0 @@
-import React from "react";
-import styled from "styled-components";
-import BuyNowPrice from "../BuyNowPrice";
-import TopBid from "../TopBid";
-
-const PriceContainerStyle = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-  justify-content: flex-end;
-  margin-top: 0.5rem;
-  flex: 1;
-`;
-
-const PriceContainer = props => {
-  const { buyNowPrice, topBid } = props;
-    return (
-    <PriceContainerStyle>
-      { buyNowPrice === null ? <></> : <BuyNowPrice buyNowPrice={buyNowPrice}/>}
-      <TopBid topBid={topBid}/>
-    </PriceContainerStyle>
-  );
-};
-
-export default PriceContainer;
